# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Auto-Healing Platform API
  description: |
    Enterprise-grade Intelligent IT Operations Self-Healing Platform API.

    ## Authentication
    All API endpoints (except `/api/v1/auth/login` and `/api/v1/auth/refresh`) require a Bearer token.

    ```
    Authorization: Bearer <access_token>
    ```

    ## Route Groups
    - **Auth** (`/api/v1/auth/*`) — Login, token refresh, user profile
    - **Common** (`/api/v1/common/*`) — Authenticated shared endpoints; tenant users may use explicit or default tenant context, platform admins only gain tenant context during Impersonation
    - **Platform** (`/api/v1/platform/*`) — Platform admin operations (users, roles, tenants, settings)
    - **Tenant** (`/api/v1/tenant/*`) — Tenant-scoped operations (plugins, CMDB, execution, healing, etc.)

  version: 1.0.0
  contact:
    name: Auto-Healing Team
    url: https://github.com/heyangguang/auto-healing
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          example: admin123456
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
        user:
          $ref: '#/components/schemas/UserBasic'
    UserBasic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        display_name:
          type: string
        email:
          type: string
        status:
          type: string
          enum: [active, inactive]
...
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
...